refactor: move disable on click to menu item base
